+++ lxconsole/api/account.py
@@ -1,14 +1,13 @@
-from flask import jsonify, request
+from flask import jsonify, request, session
 import json
 import re
 import pyotp
 from lxconsole import db, bcrypt
 from lxconsole.models import User, UserGroup, Group, TOTP, Setting
-from flask_login import login_required
+from flask_login import login_required, current_user
 from lxconsole.api.access_controls import privilege_check
 
-
-def retrieve_user_groups(user_id):
+def retrieve_account_groups(user_id):
   usergroups = UserGroup.query.filter_by(user_id=user_id).all()
   groups = []
   for usergroup in usergroups:
@@ -18,72 +17,13 @@ def retrieve_user_groups(user_id):
 
 
 @login_required
-def api_users_endpoint(endpoint):
+def api_account_endpoint(endpoint):
 
   if not privilege_check(endpoint):
     return jsonify({'data': [], 'metadata':[], 'error': 'not authorized', 'error_code': 403})
 
-
-  if endpoint == 'add_user':
-    username = request.form.get('username')
-    email = request.form.get('email')
-    first_name = request.form.get('first_name')
-    last_name = request.form.get('last_name')
-
-    #Check for password requirements
-    minimumCharacterLength = Setting.query.filter_by(name='minimumCharacterLength').first()
-    if minimumCharacterLength and int(minimumCharacterLength.value or 0) > 0:
-      if (len(request.form.get('password')) < int(minimumCharacterLength.value)):
-        password_check = False
-        alert = "Passwords must be at least " + minimumCharacterLength.value + " characters long"
-        return jsonify({"alert": alert})
-
-    enablePasswordComplexity = Setting.query.filter_by(name='enablePasswordComplexity').first()
-    if enablePasswordComplexity and enablePasswordComplexity.value == 'true':
-      requireUppercaseCharacters = Setting.query.filter_by(name='requireUppercaseCharacters').first()
-      requireLowercaseCharacters = Setting.query.filter_by(name='requireLowercaseCharacters').first()
-      requireNumbers = Setting.query.filter_by(name='requireNumbers').first()
-      requireSpecialCharacters = Setting.query.filter_by(name='requireSpecialCharacters').first()
-
-      if (len(request.form.get('password')) < 8):
-        password_check = False
-        alert = "Passwords must be at least 8 characters when complexity is enabled"
-        return jsonify({"alert": alert})
-      if requireLowercaseCharacters and requireLowercaseCharacters.value == 'true' and not re.search("[a-z]", request.form.get('password')):
-        password_check = False
-        alert = "Password complexity requires lowercase letters"
-        return jsonify({"alert": alert})
-      if requireUppercaseCharacters and requireUppercaseCharacters.value == 'true' and not re.search("[A-Z]", request.form.get('password')):
-        password_check = False
-        alert = "Password complexity requires uppercase letters"
-        return jsonify({"alert": alert})
-      if requireNumbers and requireNumbers.value == 'true' and not re.search("[0-9]", request.form.get('password')):
-        password_check = False
-        alert = "Password complexity requires numbers"
-        return jsonify({"alert": alert})
-      if requireSpecialCharacters and requireSpecialCharacters.value == 'true' and not re.search(r'[!@#$%^&*(),.?":{}|<>]', request.form.get('password')):
-        password_check = False
-        alert = "Password complexity requires special characters"
-        return jsonify({"alert": alert})
-
-    hashed_password = bcrypt.generate_password_hash(request.form.get('password')).decode('utf-8')
-    user = User(username=username, email=email, password=hashed_password, first_name=first_name, last_name=last_name)
-    db.session.add(user)
-    db.session.commit()
-    db.session.flush()
-    user_id = user.id
-
-    #Add user to group
-    usergroup = UserGroup(user_id=user_id, group_id=request.form.get('group_id'))
-    db.session.add(usergroup)
-    db.session.commit()
-
-    json_object = json.loads('{"status": 200, "alert": "New user account created"}')
-    return jsonify(json_object)
-
-
-  if endpoint == 'get_user':
-    id = request.args.get('id')
+  if endpoint == 'get_account':
+    id = current_user.id
     user = User.query.filter_by(id=id).first()
     data = {}
     data.update({'id': user.id})
@@ -91,62 +31,26 @@ def api_users_endpoint(endpoint):
     data.update({'email': user.email})
     data.update({'first_name': user.first_name})
     data.update({'last_name': user.last_name})
-    data.update({'groups': retrieve_user_groups(id)})
+    data.update({'groups': retrieve_account_groups(id)})
     data.update({'created_at': user.created_at})
     return jsonify({"metadata": data})
 
-  # Currently not used
-  if endpoint == 'get_user_mfa_status':
-    id = request.args.get('id')
+
+  if endpoint == 'get_account_mfa_status':
+    id = current_user.id
     mfa = TOTP.query.filter_by(user_id=id).first()
     return jsonify({"enabled": mfa.enabled})
 
 
-  if endpoint == 'delete_user':
-    id = request.form.get('id')
-    user = User.query.filter_by(id=id).first()
-
-    # Delete user+group relationships
-    UserGroup.query.filter_by(user_id=id).delete()
-    db.session.commit()
-
-    # Delete user TOTP information
-    TOTP.query.filter_by(user_id=id).delete()
-    db.session.commit()
-
-    # Delete user
-    db.session.delete(user)
-    db.session.commit()
-    json_object = json.loads('{"status": 200}')
-    return jsonify(json_object)
-
-
-  if endpoint == 'list_users':
-    if privilege_check('list_users_all'):
-      users = User.query.all()
-      data = []
-      for user in users:
-        groups = retrieve_user_groups(user.id)
-        mfa = TOTP.query.filter_by(user_id=user.id).first()
-        if mfa:
-          mfaEnabled = mfa.enabled
-        else:
-          mfaEnabled = False
-        data.append(dict(id=user.id, username=user.username, email=user.email, first_name=user.first_name, last_name=user.last_name, groups=groups, mfa=mfaEnabled))
-    else:
-      data = []
-    return jsonify({"data": data})
-
-
   # Currently not used
-  if endpoint == 'list_user_groups':
-    id = request.args.get('id')
-    groups = retrieve_user_groups(id)
+  if endpoint == 'list_account_groups':
+    id = current_user.id
+    groups = retrieve_account_groups(id)
     return jsonify({"data": groups})
 
 
-  if endpoint == 'update_user':
-    id = request.form.get('id')
+  if endpoint == 'update_account':
+    id = current_user.id
     user = User.query.filter_by(id=id).first()
 
     #Action for adding or removing a group or updating otp or updating password
@@ -222,8 +126,8 @@ def api_users_endpoint(endpoint):
         return jsonify({"alert": "Account profile updated"})
 
 
-  if endpoint == 'update_user_groups':
-    id = request.form.get('id')
+  if endpoint == 'update_account_groups':
+    id = current_user.id
     user = User.query.filter_by(id=id).first()
 
     #Action for adding or removing a group